Dynamic extra shortcut (#23)
If a user selects a thumb, we exclude the "Kerro lisää" shortcut and
hide thumbs from the next assistant message.

Added also two new shortcuts: "Mitä muuta kannatat?" and "Miksi?". Now
one of the shortcuts is shown randomly among the suggestions (previously
there was always "Kerro lisää").

## Future improvements

Random shortcut selection is not optimal. We could e.g. analyze the
latest message that on the server-side instead.