Update party documents (#38)
- Created new Drive folder called partyFiles2, which contains consistant
alue- and kuntavaali programs for each party.

**Discussion:**
This resulted in better and more clear files, and it seemed that the
language model had an easier time building more consistent answers
between the parties. We could consider whether this should be done for
other files as well. For example, the wikipedia articles have irrelevant
text and links, and some pdf-files are structured in a way that might be
confusing for the language model. However, it's a laborious (and manual)
task.