Update files (#39)
- Created doc files manually to remove noise from inconsistent
pdf-documents.
- Updated stores for each party

Some of the documents (e.g. RKP) are still original. They seem to work
decently, but can be written to a doc file similar to the others.

When testing, the chatbot has still hard time to find knowledge and it
might depend on very specific phrasing e.g. "jäsenmäärä" vs "jäsenet"

I'll merge this so that problems could arise sooner and indicate whether
to take a step back.